+++ changelog.md
@@ -4,7 +4,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 
 ## [Unreleased]
 
-* feat: ensure minimum value of zero for session timeout
+* feat: ensure minimum value of zero for session timeout configuration
 
 ## [3.2.0] - 2018-08-10