+++ changelog.md
@@ -7,6 +7,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * feat: ensure minimum value of zero for session timeout configuration
 * feat: show warning/success message when saving estimation to issue
 * docs: add inline documentation to CSS code
+* feat: improve error message when saving estimation to issue
 
 ## [3.2.0] - 2018-08-10